renamed search reporter file and split implementation in cpp
